+++ appinventor/misc/chatproxy/src/titan.rs
@@ -72,7 +72,7 @@ impl Converse for TitanConversation {
     }
 
     // Stub Implementation
-    fn token_count(&self, _response: &str) -> Result<i32, Box<dyn Error>> {
+    fn token_count(&self, _response: &str) -> Result<i64, Box<dyn Error>> {
         Ok(0)
     }
 }
